+++ backend/src/main/java/eu/zuinnote/example/springwebdemo/configuration/application/https/HttpsConfig.java
@@ -4,7 +4,16 @@ import org.springframework.validation.annotation.Validated;
 
 @Validated
 public class HttpsConfig {
+    private boolean disableRedirectHttps = false;
     private HttpsHeaderConfig headers;
+
+    public boolean getDisableRedirectHttps() {
+        return disableRedirectHttps;
+    }
+
+    public void setDisableRedirectHttps(boolean disableRedirectHttps) {
+        this.disableRedirectHttps = disableRedirectHttps;
+    }
 
     public HttpsHeaderConfig getHeaders() {
         return headers;